Add rule ECS0007 (#42)
Related: https://github.com/BillWagner/EffectiveCSharpAnalyzers/issues/7
Closes #32

The problem addressed by this rule is the use of methods that should
involve callback mechanisms without employing delegates for this
purpose. Specifically, methods are being invoked directly without the
use of delegate types such as `Predicate<T>`, `Action<T>`, or `Func<T>`,
which are designed to handle callbacks in a type-safe and consistent
manner.